+++ packages/dashboard/src/main.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
+import { BrowserRouter, HashRouter } from 'react-router-dom';
 import { App } from './App';
+import { IS_PUBLIC_SITE } from './lib/publicSite';
 import './index.css';
 
 class ErrorBoundary extends React.Component<
@@ -31,12 +32,14 @@ class ErrorBoundary extends React.Component<
   }
 }
 
+const Router = IS_PUBLIC_SITE ? HashRouter : BrowserRouter;
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ErrorBoundary>
-      <BrowserRouter>
+      <Router>
         <App />
-      </BrowserRouter>
+      </Router>
     </ErrorBoundary>
   </React.StrictMode>,
 );